Add /api/docs/:docId/wipeCache endpoint
This endpoint is the corollary of .../flush: it discards any local copy (the cache) so
the doc worker is forced to fetch the document and metadata from the remote storage.

Especially useful when changes has been brought to the remote doc storage
and we want to force the doc worker to fetch the version from there.
WARNING: it discard any changes that had not been pushed yet.